move filters inside a specific file
